% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadict-package.R
\docType{package}
\name{datadict-package}
\alias{datadict}
\alias{datadict-package}
\title{datadict: Validate Data Against a 'data-dict.yaml' Data Dictionary}
\description{
A data dictionary describes a collection of related tables: their columns, types, constraints, relationships, and glossary. Validate a dataset against its dictionary and read the findings as an HTML report. Reference: \url{https://data-dict.tidyverse.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://data-dict.tidyverse.org}
  \item \url{https://github.com/tidyverse/data-dict}
  \item Report bugs at \url{https://github.com/tidyverse/data-dict/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{gabor@posit.co} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID})

Authors:
\itemize{
  \item Gábor Csárdi \email{gabor@posit.co} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID})
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
